Add support for GET v3 target state
This change updates types and database format in order to allow
receiving the new format of the target state from the cloud and allow
applications to keep working.

This change has no effect on the containers as the new metadata is not
yet stored in the containers. Next commit will implement target state
PATCH and that will update the containers and user current state to
report to the v3 PATCH endpoint.